ADD: 아는 사람 찾기 컴포넌트 구현
