+++ public/data/profile/profileData.json
@@ -1,5 +1,5 @@
 {
-  "profileData": {
+  "PROFILE_DATA": {
     "userProfileUrl": "",
     "backgroundImg": "",
     "firstName": "김",
