+++ public/data/profile/profileDataUrl.json
@@ -1,5 +1,5 @@
 {
-  "profileData": {
+  "PROFILE_DATA": {
     "userProfileUrl": "https://img1.daumcdn.net/thumb/C100x100/?scode=mtistory2&fname=https%3A%2F%2Ftistory1.daumcdn.net%2Ftistory%2Fuser%2F4988582%2Fprofile%2FprofileImg%3Fv%3D1627002856736",
     "backgroundImg": "https://images.unsplash.com/photo-1633105096371-b70eb1351b9d?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1171&q=80",
     "firstName": "김",
