+++ src/pages/Profile/DropDown/sectionsDropDown.js
@@ -1,12 +1,13 @@
 import styled from 'styled-components';
 
-import { interests } from '../../data/dropDownData';
+import { sections } from '../data/dropDownData';
 
-export default function InterestsDropDown(props) {
-  const { showInterestsDropDown } = props;
+export default function SectionsDropDown(props) {
+  const { showSectionsDropDown } = props;
+
   return (
-    <StyledDropDown showInterestsDropDown={showInterestsDropDown}>
-      {interests.map(list => {
+    <StyledDropDown showSectionsDropDown={showSectionsDropDown}>
+      {sections.map(list => {
         return (
           <li>
             <div>{`${list.title}`}</div>
@@ -21,17 +22,17 @@ export default function InterestsDropDown(props) {
 const StyledDropDown = styled.ul`
   position: absolute;
   top: 60px;
-  left: 10px;
-  width: 170px;
+  left: 95px;
+  width: 200px;
   background-color: whitesmoke;
   box-shadow: 1px 2px 2px;
   border: 1px solid lightgray;
   border-radius: 5px;
   z-index: 1000;
 
-  transform: ${({ showInterestsDropDown }) =>
-    showInterestsDropDown ? 'translateY(-20px)' : 'none'};
-  opacity: ${({ showInterestsDropDown }) => (showInterestsDropDown ? 1 : 0)};
+  transform: ${({ showSectionsDropDown }) =>
+    showSectionsDropDown ? 'translateY(-20px)' : 'none'};
+  opacity: ${({ showSectionsDropDown }) => (showSectionsDropDown ? 1 : 0)};
   transition: all 0.5s ease;
 
   li {
