+++ src/pages/Profile/Modals/BasicEditModal.js
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 import styled from 'styled-components';
-import Button from '../../../../components/Button/Button';
+import Button from '../../../components/Button/Button';
 import TextInput from './components/TextInput';
 import CheckBox from './components/CheckBox';
 import AskCancel from './components/AskCancelModal';
-import theme from '../../../../styles/theme';
+import theme from '../../../styles/theme';
 
 export default function BasicEditModal(props) {
   const [showAskCancelModal, setShowAskCancelModal] = useState(false);
