+++ src/pages/Profile/Modals/CareerEditModal.js
@@ -1,16 +1,24 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import EndInput from './components/EndInput';
 import StartInput from './components/StartInput';
-import Button from '../../../../components/Button/Button';
+import Button from '../../../components/Button/Button';
 import TextInput from './components/TextInput';
+import CheckBox from './components/CheckBox';
 import TextArea from './components/TextArea';
 import AskCancel from './components/AskCancelModal';
-import theme from '../../../../styles/theme';
+import theme from '../../../styles/theme';
 
-export default function EducationEditModal(props) {
+export default function CareerEditModal(props) {
+  const [isWorkingNow, setIsWorkingNow] = useState(true);
+  const [isPositionClosed, setIsPositionClosed] = useState(false);
   const [showAskCancelModal, setShowAskCancelModal] = useState(false);
   const [hasChanged, setHasChanged] = useState(false);
+
+  useEffect(() => {
+    // 현재 직책 종료 여부 초기화
+    if (!workingCurrently) setIsWorkingNow(workingCurrently);
+  }, []);
 
   const openAskCancelModal = e => {
     e.preventDefault();
@@ -23,98 +31,145 @@ export default function EducationEditModal(props) {
     setShowAskCancelModal(false);
   };
 
+  const toggleIsWorkingNow = () => {
+    setIsWorkingNow(!isWorkingNow);
+  };
+
+  const toggleIsPositionClosed = () => {
+    setIsPositionClosed(!isPositionClosed);
+  };
+
   const toggleHasChanged = () => {
     setHasChanged(true);
   };
 
-  const { showEducationEditModal, closeEducationEditModal } = props;
-  const { schoolName, degree, major, startDate, endDate, gpa, activity, desc } =
-    props.education || {};
+  const { selectedCareer, showCareerEditModal, closeCareerEditModal } = props;
+  const {
+    position,
+    employmentType,
+    companyName,
+    country,
+    city,
+    isWorkingNow: workingCurrently,
+    startDate,
+    endDate,
+    industry,
+    desc,
+  } = props.selectedCareer || {};
 
   return (
-    <StyledEducationEditModal showEducationEditModal={showEducationEditModal}>
-      {showEducationEditModal && (
+    <StyledCareerAddModal showCareerEditModal={showCareerEditModal}>
+      {showCareerEditModal && (
         <div>
           <DimBg
             scrollY={window.scrollY}
-            onClick={hasChanged ? openAskCancelModal : closeEducationEditModal}
+            onClick={hasChanged ? openAskCancelModal : closeCareerEditModal}
           />
           <Container scrollY={window.scrollY}>
             <BoxHeader>
-              <h2>학력사항 수정</h2>
+              <h2>경력 사항 수정</h2>
               <div className="highlightCircle">
                 <i
                   className="fal fa-times"
                   onClick={
-                    hasChanged ? openAskCancelModal : closeEducationEditModal
+                    hasChanged ? openAskCancelModal : closeCareerEditModal
                   }
                 />
               </div>
             </BoxHeader>
             <BoxBody onChange={toggleHasChanged}>
               <form method="POST">
-                <SchoolInput>
+                <PositionInput>
                   <TextInput
-                    title="학교"
-                    name="schoolName"
-                    defaultValue={schoolName}
+                    title="직책"
+                    name="position"
+                    defaultValue={position}
                     textLimit={100}
                     warningText="100자를 초과할 수 없습니다."
                     isNullable={false}
                   />
-                </SchoolInput>
-                <DegreeInput>
+                </PositionInput>
+                <EmploymentTypeInput>
                   <TextInput
-                    title="학위"
-                    name="degree"
-                    defaultValue={degree}
+                    title="고용형태"
+                    name="employmentType"
+                    defaultValue={employmentType}
                     textLimit={100}
+                    warningText="100자를 초과할 수 없습니다."
+                    isNullable={false}
+                  />
+                  <EmploymentMoreInfo>국가별 고용형태</EmploymentMoreInfo>
+                  <a
+                    href="https://www.linkedin.com/help/linkedin/answer/123304"
+                    target="_blank"
+                    rel="noreferrer"
+                  >
+                    자세히 보기
+                  </a>
+                </EmploymentTypeInput>
+                <CompanyNameInput>
+                  <TextInput
+                    title="회사이름"
+                    name="companyName"
+                    defaultValue={companyName}
+                    textLimit={100}
+                    warningText="100자를 초과할 수 없습니다."
+                    isNullable={false}
+                  />
+                </CompanyNameInput>
+                <RegionInput>
+                  <TextInput
+                    title="지역"
+                    name="region"
+                    textLimit={100}
+                    defaultValue={`${country} ${city}`}
                     warningText="100자를 초과할 수 없습니다."
                     isNullable={true}
                   />
-                </DegreeInput>
-                <MajorInput>
-                  <TextInput
-                    title="전공"
-                    name="major"
-                    defaultValue={major}
-                    textLimit={100}
-                    warningText="100자를 초과할 수 없습니다."
-                    isNullable={true}
+                </RegionInput>
+                <CheckBoxWrapper isWorkingNow={isWorkingNow}>
+                  <CheckBox
+                    id="cbIsWorking"
+                    name="isWorkingNow"
+                    state={isWorkingNow}
+                    text="현재 이 업무로 근무 중"
+                    onClick={toggleIsWorkingNow}
                   />
-                </MajorInput>
+                </CheckBoxWrapper>
                 <StartAndEndWrapper>
-                  <StartInput title="입학일" defaultValue={startDate} />
-                  <EndInput title="졸업일(예정)" defaultValue={endDate} />
+                  <StartInput title="시작일" defaultValue={startDate} />
+                  <EndInput
+                    title="종료일"
+                    defaultValue={endDate}
+                    state={isWorkingNow}
+                  />
                 </StartAndEndWrapper>
-                <GradeInput>
+                {isWorkingNow && (
+                  <CheckBoxWrapper isPositionClosed={isPositionClosed}>
+                    <CheckBox
+                      id="cbPositionClosed"
+                      name="isPositionClosed"
+                      state={isPositionClosed}
+                      text={`현재 직책 종료 - ${selectedCareer.position} at ${selectedCareer.companyName}`}
+                      onClick={toggleIsPositionClosed}
+                    />
+                  </CheckBoxWrapper>
+                )}
+                <IndustryInput>
                   <TextInput
-                    title="학점"
-                    name="grade"
-                    defaultValue={gpa}
-                    textLimit={10}
-                    warningText="10자를 초과할 수 없습니다."
-                    isNullable={false}
+                    title="업계"
+                    name="industry"
+                    defaultValue={industry}
                   />
-                </GradeInput>
-                <ActivityInput>
-                  <TextArea
-                    title="동아리나 학회"
-                    name="activity"
-                    defaultValue={activity}
-                    rows={3}
-                    textLimit={500}
-                    warningText="동아리나 학회는 500자를 넘을 수 없습니다."
-                  />
-                </ActivityInput>
+                </IndustryInput>
                 <DescInput>
                   <TextArea
                     title="설명"
                     name="description"
-                    defaultValue={desc}
                     rows={3}
-                    textLimit={1000}
-                    warningText="설명은 1,000자를 넘을 수 없습니다."
+                    defaultValue={desc}
+                    textLimit={2000}
+                    warningText="설명은 2,000자를 넘을 수 없습니다."
                   />
                 </DescInput>
               </form>
@@ -140,24 +195,23 @@ export default function EducationEditModal(props) {
                   bgc={theme.colors.primary}
                   color={theme.colors.white}
                   text="저장"
-                  onClick={closeEducationEditModal}
+                  onClick={closeCareerEditModal}
                 />
               </SaveBtnWrapper>
             </BoxTail>
           </Container>
-
           <AskCancel
             state={showAskCancelModal}
             closeAskCancelModal={closeAskCancelModal}
-            closeUpperModal={closeEducationEditModal}
+            closeUpperModal={closeCareerEditModal}
           />
         </div>
       )}
-    </StyledEducationEditModal>
+    </StyledCareerAddModal>
   );
 }
 
-const StyledEducationEditModal = styled.div``;
+const StyledCareerAddModal = styled.div``;
 
 const DimBg = styled.div`
   position: absolute;
@@ -224,28 +278,46 @@ const BoxBody = styled.div`
   overflow-x: hidden;
 `;
 
-const SchoolInput = styled.div`
+const PositionInput = styled.div`
   margin-top: 10px;
 `;
 
-const DegreeInput = styled.div`
-  margin-top: 30px;
+const EmploymentTypeInput = styled.div`
+  margin-top: 40px;
+  color: grey;
+  font-size: 15px;
+
+  a {
+    font-size: 16px;
+  }
 `;
 
-const MajorInput = styled.div`
-  margin-top: 30px;
+const EmploymentMoreInfo = styled.div`
+  margin-top: 10px;
+
+  a {
+    margin-top: 5px;
+  }
+`;
+
+const CompanyNameInput = styled.div`
+  margin-top: 40px;
+`;
+
+const RegionInput = styled.div`
+  margin-top: 40px;
+`;
+
+const CheckBoxWrapper = styled.div`
+  margin: 40px 0 20px 0;
 `;
 
 const StartAndEndWrapper = styled.div`
   margin-top: 30px;
 `;
 
-const GradeInput = styled.div`
-  margin-top: 30px;
-`;
-
-const ActivityInput = styled.div`
-  margin-top: 30px;
+const IndustryInput = styled.div`
+  margin-top: 40px;
 `;
 
 const DescInput = styled.div`
